remove unused weird code
