+++ src/app.html
@@ -43,7 +43,6 @@
     <link rel="preload" href="/hero.svg" as="image" type="image/svg+xml" />
     <link rel="preload" href="/keyboard.svg" as="image" type="image/svg+xml" />
     <link rel="preload" href="/Screen.svg" as="image" type="image/svg+xml" />
-    <link rel="preload" href="/honeymilk.svg" as="image" type="image/svg+xml" />
     <link
       rel="preload"
       href="/magic-hand.svg"
