+++ src/routes/components/Description.svelte
@@ -63,68 +63,4 @@
       width: 100%;
     }
   }
-
-  .cards {
-    display: flex;
-    gap: 2rem;
-    max-width: 100%;
-    justify-content: center;
-    align-items: center;
-    margin-block-start: calc(-1 * var(--offset));
-    padding-block: 3rem;
-    position: relative;
-    div {
-      overflow: hidden;
-      position: relative;
-      height: auto;
-      max-width: 300px;
-      aspect-ratio: 5/7;
-      box-shadow:
-        0px 10px 16px -3px oklch(0.6 0.01 45 / 0.1),
-        0px 0px 30px 0px oklch(0.6 0.01 45 / 0.1),
-        0px 8px 7px 0px oklch(0.6 0.01 45 / 0.1);
-      &::before {
-        content: "";
-        position: absolute;
-      }
-      border: 2px solid white;
-      border-radius: 1rem;
-      transition-property: transform, rotate;
-      transition-duration: 600ms;
-      /* cooked overshoot */
-      transition-timing-function: cubic-bezier(0, 0.74, 0.2, 1.6);
-      img {
-        position: relative;
-        display: block;
-        object-fit: cover;
-      }
-
-      &:nth-child(1) {
-        background: var(--purple);
-        rotate: -7.19deg;
-        transform: translate(10px, 20px);
-      }
-      &:nth-child(2) {
-        background: white;
-        rotate: 7.89deg;
-        transform: translate(0px, -8px);
-        z-index: 1;
-      }
-      &:nth-child(3) {
-        background: var(--yellow);
-        rotate: -6.37deg;
-        transform: translate(10px, -12px);
-      }
-      &:nth-child(4) {
-        background: black;
-        rotate: 8.93deg;
-        transform: translate(-30px, -10px);
-        z-index: 1;
-      }
-    }
-    &:hover div {
-      transform: translate(0);
-      rotate: 0deg;
-    }
-  }
 </style>
