--- src/routes/components/features/available-today.md
@@ -1,7 +1,0 @@
-## Available today
-
-- **Snazzy personal website:** Link to all your stuff and begin reconsolidating your fragmented online identity, [POSSE](https://indieweb.org/POSSE) style.
-- **Custom domain:** Get a sub-domain on `weird.one`, or set up your own personal domain.
-- Automatic Linktree importer (has to be manually requested for now)
-- A social network without a content-creation imperative
-- Big dreams ✨
