--- src/routes/components/features/available-today.svelte
@@ -1,11 +1,0 @@
-<script>
-  import Card from "./card.svelte";
-  import rawMarkdown from "./available-today.md?raw";
-  import Markdown from "$lib/Markdown.svelte";
-</script>
-
-<Card>
-  <span slot="icon">➢</span>
-
-  <Markdown {rawMarkdown} slot="body" />
-</Card>
