--- src/routes/components/features/coming-soon.md
@@ -1,8 +1,0 @@
-## Coming soon
-
-- Lots more importers (taking requests!)
-- [Custom themes](https://github.com/muni-town/weird/issues/11)
-- [Collaborative writing (CRDT)](https://github.com/muni-town/weird/issues/264)
-- [Login with Bluesky (ATProto)](https://github.com/muni-town/weird/issues/206)
-- [Private content](https://blog.muni.town/capabilities-and-identity-with-leaf/)
-- [Teams and Projects](https://github.com/muni-town/weird/discussions/20) (👈this is where the socializing begins in earnest)
