--- src/routes/components/features/coming-soon.svelte
@@ -1,11 +1,0 @@
-<script>
-  import Card from "./card.svelte";
-  import rawMarkdown from "./coming-soon.md?raw";
-  import Markdown from "$lib/Markdown.svelte";
-</script>
-
-<Card>
-  <span slot="icon">➢</span>
-
-  <Markdown {rawMarkdown} slot="body" />
-</Card>
