--- src/routes/components/features/magic-website-creator.md
@@ -1,7 +1,0 @@
-To fully inhabit the World Wide Web you need to create your digital self within it.
-
-For a lot of us, the “simple” prospect of filling in that web page with content presents an insurmountable [writer’s block](https://en.wikipedia.org/wiki/Writer%27s_block).
-
-That’s why we start with _link lists_ as a primitive building block. Make a _list_ of 1, 3 or however many `links` that say something, anything, about you.
-
-Whatever you’ve listed, you’ve done it! You told us something about your self, and with that you are taking up a bit of space on the web, as you should.
