--- src/routes/components/features/web-passports.md
@@ -1,3 +1,0 @@
-Weird is the “Google-login” button of the alt-web. A [web of the people](https://blog.erlend.sh/web-of-the-people), powered by people's websites instead of those pesky totalitarian mega-platforms.
-
-By making - or merely connecting - your website with Weird you will soon be able to access and interact with the small, open and indie web just as easily as the regular web works today. Same conveniences, without the dark patterns, spying and data theft.
