--- src/routes/components/features/web-passports.svelte
@@ -1,12 +1,0 @@
-<script>
-  import Card from "./card.svelte";
-  import rawMarkdown from "./web-passports.md?raw";
-  import Markdown from "$lib/Markdown.svelte";
-</script>
-
-<Card>
-  <span slot="icon">➣</span>
-  <span slot="title">Web Passports</span>
-
-  <Markdown {rawMarkdown} slot="body" />
-</Card>
